// other/src/shared/types/interfaces.ts
export interface SendResponseParams {
  res: Response;
  data?: any;
  message?: string;
  statusCode?: number;
}

export interface SendErrorParams {
  res: Response;
  message?: string;
  statusCode?: number;
  errors?: IError[];
}
// ...
export interface IFindOptions<T> extends FindManyOptions<T> {
  page?: number;
  size?: number;
  keyword?: string; // Keyword for text search
  searchFields?: (keyof T)[]; // Fields to search in
  timeField?: keyof T; // Field to apply the date range filter on
  summaryFields?: (keyof T)[]; // Fields to summarize
  type?: string; // Example field for filtering by type
  status?: string; // Example field for filtering by status
  startAt?: Date; // Example field for filtering by date range
  endAt?: Date; // Example field for filtering by date range
  sortBy?: string; // Field to sort by
  sortOrder?: "ASC" | "DESC"; // Sort direction
  isFinished?: boolean; // Example field for filtering by completion status
  filterOptions?: (keyof T)[]; // Additional filter options
  projectId?: string; // Example field for filtering by project ID
  employeeId?: string; // Example field for filtering by employee ID
  fundId?: string; // Example field for filtering by fund ID
  advanceId?: string; // Example field for filtering by advance ID

  productIds?: string[]; // Example field for filtering by product IDs
  storeIds?: string[]; // Example field for filtering by store IDs
  employeeIds?: string[]; // Example field for filtering by employee IDs
  fundCategoryIds?: string[]; // Example field for filtering by fundCategory IDs
  partnerIds?: string[]; // Example field for filtering by partner IDs
  toFundIds?: string[]; // Example field for filtering by fund IDs
}